feat: Add scripts for comparing and preparing repository versions
The new files `compare.sh` and `prepare.sh` are added to the repository. These scripts are used to compare the local version with the remote version and set the `SYNCED` environment variable accordingly. The `prepare.sh` script also downloads the necessary files from the latest version of the `linuxserver/docker-duplicati` repository and updates the `Dockerfile` with the `WEBUI` environment variable.