feat: Update prepare.sh to dynamically set the webservice port
This commit modifies the prepare.sh script to dynamically set the webservice port in the file "root/etc/s6-overlay/s6-rc.d/svc-duplicati/run". It replaces the hardcoded port number 8200 with the value of the WEBUI environment variable. Additionally, it adds a new option "--webservice-port=${WEBUI}" to the same file. These changes improve the flexibility and configurability of the Docker container.